pi/requirements.txt
supabase==2.7.4
gpiozero==2.0.1
python-dotenv==1.0.1
google-genai==0.3.0
opencv-python-headless==4.10.0.84
# lgpio is installed via apt (python3-lgpio) — see README
